Move stable preprocess test closer to implementation
